// src/lib/types/movieType.ts
export interface Movie {
    adult: boolean;
    backdrop_path: string;
    genre_ids: number[];
    id: number;
    original_language: string;
    original_title: string;
    overview: string;
    popularity: number;
    poster_path: string;
    release_date: string;
    title: string;
    video: boolean;
    vote_average: number;
    vote_count: number;
  }
  
  export interface MovieApiResponse {
    page: number;
    results: Movie[];
    total_pages: number;
    total_results: number;
  }


export interface MovieDetails {
  id: number
  title: string
  overview: string
  poster_path: string
  release_date: string
  genres: { id: number; name: string }[]
  vote_average: number
  runtime: number
}

export interface MovieDetailsPageProps {
  movie: MovieDetails | null
  error?: string
}
  